# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Contoh API documentation",
        "description": "Contoh API doc menggunakan OpenAPI/Swagger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "nadzirafarahiya@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "paths": {
        "/api/gallery": {
            "get": {
                "tags": [
                    "gallery"
                ],
                "summary": "Returns a Sample API response",
                "description": "A Sample API response to test out the API",
                "operationId": "gallery",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the data",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Successfully retrieved the data",
                                    "gallery": {
                                        "id": 1,
                                        "title": "Gallery PPW2",
                                        "description": "Gallery PPW2",
                                        "picture": "https://via.placeholder.com/150",
                                        "created_at": "2024-11-10T07:00:00.000000Z",
                                        "updated_at": "2024-11-10T07:00:00.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "detail": "No gallery data found"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/greet": {
            "get": {
                "tags": [
                    "greeting"
                ],
                "summary": "Returns a Sample API response",
                "description": "A Sample API response to test out the API",
                "operationId": "greet",
                "parameters": [
                    {
                        "name": "firstname",
                        "in": "query",
                        "description": "First name of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lastname",
                        "in": "query",
                        "description": "Last name of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Successful operation"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "gallery",
            "description": "gallery"
        },
        {
            "name": "greeting",
            "description": "greeting"
        }
    ]
}